Changed how the decides to hunt (and other fixes)
* The AI now decides to hunt based on how long it has been since they
last spotted an Animal. With every update since they haven't seen an
animal the chance that they will hunt when hungry (or have nothing todo)
increases. This also means that the AI will hunt before completing
buildnings.

* Started fixing how pathfinding is used by the AI. Fixed a few states
that may sometimes save a faulty path. To follow up on this the
pathStack needs to be looked into and reworked for a bugfree movingState

* Expressions corrected