Updated AIs gather thinking
* The AI will start with gathering water and then food. This gives it a
good start of the day and forces it to find a lake as soon as possible
so it will not die later on because it could not find one when
hungry/thirsty.

* The AI will now gather 3 water and at least 3 food when gathering so
it have some to spare for later. This means the NPC wont have to go back
and forth between a lake as often which gives increased productivity